refactor: split shared/mod.ts into pure and TUI-dependent barrels
`shared/mod.ts` was a contaminated barrel: it re-exported both pure
utilities (formatDuration, stripAnsi, etc.) and TUI-dependent functions
(makeUI, showInterviewRound, showNextAction, showConfirm). Any non-TUI
caller importing from mod dragged in @gsd/pi-tui transitively.

The workaround was a fire-and-forget dynamic import in ui.ts that
populated a module-level `_piTui` variable, with `piTui()` reading it
synchronously. This was a race condition that happened to work because
the ESM loader resolved fast enough. Errors were silently swallowed via
`.catch(() => {})`.

Fix: barrel split.

- `shared/tui.ts` — new public barrel re-exporting all TUI-dependent
  symbols (makeUI, UI, showInterviewRound, showNextAction, showConfirm
  and related types)
- `shared/mod.ts` — pure utilities only, zero @gsd/pi-tui dependency
- `shared/ui.ts` — static top-level import replaces lazy pattern;
  _piTui global and piTui() accessor removed
- 14 caller files updated: TUI imports move to shared/tui.js
- Contract test updated: now asserts mod.ts has no @gsd/pi-tui import
  (structural guarantee, not a text grep for a workaround)

No behaviour changes. makeUI signature unchanged. tsc clean, 27/27 tests pass.
