+++ src/resources/extensions/shared/mod.ts
@@ -1,7 +1,6 @@
 // Barrel file — re-exports consumed by external modules
 
 export {
-	makeUI,
 	GLYPH,
 	INDENT,
 	STATUS_GLYPH,
@@ -27,10 +26,6 @@ export {
 
 export { shortcutDesc } from "./terminal.js";
 export { toPosixPath } from "./path-display.js";
-export { showInterviewRound } from "./interview-ui.js";
-export type { Question, QuestionOption, RoundResult } from "./interview-ui.js";
-export { showNextAction } from "./next-action-ui.js";
-export { showConfirm } from "./confirm-ui.js";
 export { sanitizeError, maskEditorLine } from "./sanitize.js";
 export { formatDateShort, truncateWithEllipsis } from "./format-utils.js";
 export { splitFrontmatter, parseFrontmatterMap } from "./frontmatter.js";
