+++ src/resources/extensions/shared/ui.ts
@@ -29,23 +29,7 @@
  */
 
 import { type Theme } from "@gsd/pi-coding-agent";
-
-// ─── Lazy @gsd/pi-tui resolution ─────────────────────────────────────────────
-// Deferred to first makeUI() call so that importing this module (via the
-// shared/mod barrel) does not blow up when @gsd/pi-tui cannot be resolved —
-// e.g. for commands like /exit that never render TUI components.
-
-import { createRequire } from "node:module";
-
-type PiTuiFns = typeof import("@gsd/pi-tui");
-let _piTui: PiTuiFns | undefined;
-function piTui(): PiTuiFns {
-	if (!_piTui) {
-		const _require = createRequire(import.meta.url);
-		_piTui = _require("@gsd/pi-tui") as PiTuiFns;
-	}
-	return _piTui;
-}
+import { truncateToWidth, visibleWidth, wrapTextWithAnsi } from "@gsd/pi-tui";
 
 // ─── Glyphs ───────────────────────────────────────────────────────────────────
 // Change these to restyle every cursor, checkbox, and indicator at once.
@@ -217,7 +201,6 @@ export interface UI {
 export function makeUI(theme: Theme, width: number): UI {
 	// ── Internal helpers ───────────────────────────────────────────────────────
 
-	const { truncateToWidth, visibleWidth, wrapTextWithAnsi } = piTui();
 
 	const add = (s: string): string => truncateToWidth(s, width);
 	const wrap = (s: string): string[] => wrapTextWithAnsi(s, width);
