+++ src/resources/extensions/slash-commands/create-slash-command.ts
@@ -1,5 +1,5 @@
 import type { ExtensionAPI } from "@gsd/pi-coding-agent";
-import { showInterviewRound, type Question, type RoundResult } from "../shared/mod.js";
+import { showInterviewRound, type Question, type RoundResult } from "../shared/tui.js";
 
 export default function createSlashCommand(pi: ExtensionAPI) {
 	pi.registerCommand("create-slash-command", {
